Add some checks and a power saving timeout
Added a simple check for out-of-bounds position sensor readings

The power supply regulator for 5V was getting very warm if the
rotator ended travel with both az and el moving in a positive direction.
This was due to both relays for direction being energized. While
measurements showed the regulators to be operating within limits,
I added a timout (really a loop counter) that turns off the relays
if no movement is under way and no movement commands have been
sent to the controller in a while.

Signed-off-by: Steve Conklin <sconklin@canonical.com>